added necessary classes to fetch current weather

+++ app/src/main/java/soutvoid/com/DsrWeatherApp/domain/wind/Wind.kt
@@ -6,5 +6,5 @@ data class Wind(
         @SerializedName("speed")
         val speed : Double,
         @SerializedName("deg")
-        val degrees : Int
+        val degrees : Double
 )
